+++ src/apiCalls/__mocks__/movieCleaner.js
@@ -1,0 +1,11 @@
+export const movieCleaner = jest.fn().mockImplementation((moviesArray) => {
+  return moviesArray.map((movie, index) => {
+    return ({
+      key: movie.title + index,
+      title: movie.title, 
+      overview: movie.overview,
+      poster: movie.poster_path,
+      rating: movie.vote_average
+    });
+  });
+});
