+++ src/apiCalls/apiCalls.js
@@ -1,4 +1,5 @@
 import apiKey from './apiKey';
+import { movieCleaner } from './movieCleaner'
 
 export const url = `https://api.themoviedb.org/3/discover/movie?api_key=${apiKey}&primary_release_date.gte=2018-02-27&primary_release_date.lte=2018-03-27`;
 
@@ -11,18 +12,6 @@ export const getMovies = async (url) => {
   } catch (error) {
     throw new Error('Error getting movies');
   }
-};
-
-export const movieCleaner = (moviesArray) => {
-  return moviesArray.map((movie, index) => {
-    return ({
-      key: movie.title + index,
-      title: movie.title, 
-      overview: movie.overview,
-      poster: movie.poster_path,
-      rating: movie.vote_average
-    });
-  });
 };
 
 export const userLogin = async credentials => {
